issue: Agent Permission Info
This addresses an issue where when exporting Agents and an Agent doesn't
have a Permission or Permission lookp fails the system throws a fatal error.
This is due to attempting to call `getInfo()` on `getPermission()` when it
doesn't exist or returns `null`/`false`. This adds a null coalescing
operator so that if the permission lookup fails or returns invalid data we
will use an empty array to avoid fatal errors and let the export continue.